fix: the four CI failures that are not formatting
MSVC rejected `.clip = SCAV_CLIP_NONE` in a braced initializer: an unnamed C
enum holding 0xFFFFFFFF has an implementation-defined underlying type, which is
the hazard §6 bans typeless enums for. A C header cannot spell the underlying
type before C23, so the constant is cast at the one site that braces it. The
comparison sites are plain conversions and MSVC accepts those.

builder_tests.cpp used std::memcpy without <cstring>. libc++ supplies it
transitively and libstdc++ does not, so every gcc and clang-libstdcxx row broke
while macOS was green.

test_provisioning shelled out to `git check-ignore`, which exits 128 in the CI
container -- git refuses a repository it does not own, and the assertion read
that as "not ignored". No other test in the suite needs a working git, and the
build is meant to work from a plain tarball, so the claim is now a text
assertion on .gitignore.

The perf scaling ratios were computed from min-of-N on both sides, which is the
wrong estimator for a ratio under load: the shorter run finds an uncontended
window more often than the longer one, so the measured growth is biased upward.
Two macOS rows failed on two different assertions this way (14.5x against an
11.9x bar, 8.5x against 6x). Ratios now sum every run instead, so contention
inflates numerator and denominator together; the bars are untouched. Locally the
same assertions read 5.1x, 3.5x and 4.0x.

That last one is P0/P2 test infrastructure this branch does not otherwise touch,
changed because this branch is what tipped it: two more libraries and two more
test executables mean more compiles in flight while a timing test runs as a
build step.

+++ src/draw/drawlist.cpp
@@ -97,7 +97,9 @@ void push_prim(DrawList &d,
   d.prims.push_back({ .kind = kind,
                       .depth = depth,
                       .style = style,
-                      .clip = SCAV_CLIP_NONE,
+                      // Cast because a C enum's underlying type is the
+                      // implementation's choice, and MSVC calls this narrowing.
+                      .clip = static_cast<uint32_t>(SCAV_CLIP_NONE),
                       .origin_kind = static_cast<uint32_t>(origin.kind),
                       .origin_ordinal = origin.ordinal,
                       .points = points,
